Split QueryLogger into its own source file
