+++ enterprise/query-logging/src/main/java/org/neo4j/kernel/impl/query/QueryLogFormatter.java
@@ -1,0 +1,105 @@
+/*
+ * Copyright (c) 2002-2017 "Neo Technology,"
+ * Network Engine for Objects in Lund AB [http://neotechnology.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Neo4j is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.neo4j.kernel.impl.query;
+
+import java.util.Collection;
+import java.util.Collections;
+import java.util.Map;
+
+import org.neo4j.helpers.Strings;
+import org.neo4j.kernel.api.query.QuerySnapshot;
+
+class QueryLogFormatter
+{
+    static void formatPageDetails( StringBuilder result, QuerySnapshot query )
+    {
+        result.append( query.pageHits() ).append( " page hits, " );
+        result.append( query.pageFaults() ).append( " page faults - " );
+    }
+
+    static void formatAllocatedBytes( StringBuilder result, QuerySnapshot query )
+    {
+        Long bytes = query.allocatedBytes();
+        if ( bytes != null )
+        {
+            result.append( bytes ).append( " B - " );
+        }
+    }
+
+    static void formatDetailedTime( StringBuilder result, QuerySnapshot query )
+    {
+        result.append( "(planning: " ).append( query.planningTimeMillis() );
+        Long cpuTime = query.cpuTimeMillis();
+        if ( cpuTime != null )
+        {
+            result.append( ", cpu: " ).append( cpuTime );
+        }
+        result.append( ", waiting: " ).append( query.waitTimeMillis() );
+        result.append( ") - " );
+    }
+
+    static void formatMap( StringBuilder result, Map<String,Object> params )
+    {
+        formatMap( result, params, Collections.emptySet() );
+    }
+
+    static void formatMap( StringBuilder result, Map<String,Object> params, Collection<String> obfuscate )
+    {
+        result.append( '{' );
+        if ( params != null )
+        {
+            String sep = "";
+            for ( Map.Entry<String,Object> entry : params.entrySet() )
+            {
+                result
+                        .append( sep )
+                        .append( entry.getKey() )
+                        .append( ": " );
+
+                if ( obfuscate.contains( entry.getKey() ) )
+                {
+                    result.append( "******" );
+                }
+                else
+                {
+                    formatValue( result, entry.getValue() );
+                }
+                sep = ", ";
+            }
+        }
+        result.append( "}" );
+    }
+
+    private static void formatValue( StringBuilder result, Object value )
+    {
+        if ( value instanceof Map<?,?> )
+        {
+            formatMap( result, (Map<String,Object>) value );
+        }
+        else if ( value instanceof String )
+        {
+            result.append( '\'' ).append( value ).append( '\'' );
+        }
+        else
+        {
+            result.append( Strings.prettyPrint( value ) );
+        }
+    }
+}
